requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-trino==5.7.0
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
boto3<1.35
python-dotenv
requests
astronomer-cosmos
pyarrow==20.0.0
pyiceberg[glue,s3fs]==0.8.1
